working on cmake

+++ DotNetLibrary/include/StringDotNet.h
@@ -9,7 +9,7 @@ namespace DotnetLibrary
 	class DLL_EXPORT String : public Object
 	{
 		Char* characterData;
-		size_t	length;
+		Int32	length;
 		Boolean allocate;
 	public:
 		String(const Char* string);
@@ -19,7 +19,7 @@ namespace DotnetLibrary
 		virtual				~String();
 		Char*				get_Buffer() { return(characterData); }
 		Char				get_Chars(const Int32 index);
-		Int32				get_Length() { return(length); };
+		Int32				get_Length() { return((Int32)length); };
 		Char*				Address(Int32 index1, Int32 index2 = -1, Int32 index3 = -1);
 		static	Int32		Compare(Char* s1, Int32 s1Index, Char* s2, Int32 s2Index, Int32 length, Boolean ignoreCase);
 		static	Int32		Compare(String* s1, String* s2, Boolean ignoreCase = false);
